refactor(step3): order package로 이동
chore(step3): 더이상 사용하지 않는 package 삭제
