+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -1,9 +1,20 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain;
 
-import javax.persistence.*;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.ForeignKey;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import javax.persistence.Transient;
 import kitchenpos.ordertable.domain.OrderTable;
 
 @Table(name = "orders")
