+++ src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -1,8 +1,17 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain;
 
-import javax.persistence.*;
 import java.math.BigDecimal;
 import java.util.UUID;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.ForeignKey;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import javax.persistence.Transient;
 import kitchenpos.menu.domain.Menu;
 
 @Table(name = "order_line_item")
